requirement.txt
opencv-python
ultralytics
transformers
supervision
numpy
pandas
matplotlib
rich.progress
gradio